src/classes/Vacuum.java
package classes;

import enums.Orientation;

public class Vacuum {
    private Orientation orientation;

    public Vacuum(Orientation orientation) {
        this.orientation = orientation;
    }

    public void turnRight() {
        switch (orientation) {
            case N:
                orientation = Orientation.E;
                break;
            case E:
                orientation = Orientation.S;
                break;
            case S:
                orientation = Orientation.W;
                break;
            case W:
                orientation = Orientation.N;
                break;
        }
    }

    public void turnLeft() {
        switch (orientation) {
            case N:
                orientation = Orientation.W;
                break;
            case E:
                orientation = Orientation.N;
                break;
            case S:
                orientation = Orientation.E;
                break;
            case W:
                orientation = Orientation.S;
                break;
        }
    }

    public Orientation getOrientation() {
        return orientation;
    }
}
